fixes for rendering multiple diagrams in a row
